+++ CHANGES.rst
@@ -9,6 +9,9 @@ Changelog
   [gbastien]
 - Fail institution edit form validation if an iA.Delib category is mapped multiple times.
   [odelaere]
+- Adapted `SelectMeetingWidget` used for the `seances` criterion to make
+  zero count values shown and selectable.
+  [gbastien]
 
 
 1.3.3.1 (2021-08-16)